Use BigDecimal instead of double in "price" attribute of GiftCertificate

+++ service/src/main/java/com/epam/esm/service/util/impl/CertificateValidatorImpl.java
@@ -6,6 +6,7 @@ import com.epam.esm.service.exception.ServiceException;
 import com.epam.esm.service.util.CertificateValidator;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 @Component
@@ -46,7 +47,7 @@ public class CertificateValidatorImpl implements CertificateValidator {
         }
     }
 
-    private void validatePrice(double price) throws ServiceException{
+    private void validatePrice(BigDecimal price) throws ServiceException{
         if (price < 0) {
             throw new ServiceException("Failed to validate: certificate price is negative",
                     ErrorCodeEnum.CERTIFICATE_VALIDATION_ERROR);
